CSS Refactor Phase 2
So this is the second stage of the css refactoring started last year. It
basically refines what was done in phase 1; declaring variables,
separating the rules which affect the layout (essentially the
responsiveness of the website) and how it looks.

The outward look of the site should still be intact, though there may be
a few minor tweaks - mostly for consistency. Most of the changes are
with the internals (hopefully) making things more logical and easier to
change and maintain.

The next and final (major) refactoring step is to implement "themes" for
the site. Where people will be able to design their own version of the
code club website and contribute it to the repository. Then hopefully
there is a way to have visitors come to the site and choose which theme
is active.